fix indentation, split 386EX into separate .cpp/.h files and remove FETCH override
