+++ scripts/src/cpu.lua
@@ -1156,6 +1156,8 @@ if CPUS["I386"] then
 	files {
 		MAME_DIR .. "src/devices/cpu/i386/i386.cpp",
 		MAME_DIR .. "src/devices/cpu/i386/i386.h",
+		MAME_DIR .. "src/devices/cpu/i386/i386ex.cpp",
+		MAME_DIR .. "src/devices/cpu/i386/i386ex.h",
 		MAME_DIR .. "src/devices/cpu/i386/athlon.cpp",
 		MAME_DIR .. "src/devices/cpu/i386/athlon.h",
 		MAME_DIR .. "src/devices/cpu/i386/cache.h",
