+++ src/devices/cpu/i386/i386.h
@@ -1,5 +1,5 @@
 // license:BSD-3-Clause
-// copyright-holders:Ville Linde, Barry Rodewald, Carl, Philip Bennett, Felipe Sanches
+// copyright-holders:Ville Linde, Barry Rodewald, Carl, Philip Bennett
 #ifndef MAME_CPU_I386_I386_H
 #define MAME_CPU_I386_I386_H
 
@@ -13,10 +13,8 @@
 #endif
 
 #include "divtlb.h"
+
 #include "i386dasm.h"
-#include "machine/pic8259.h"
-#include "machine/pit8253.h"
-#include "machine/ins8250.h"
 
 #define INPUT_LINE_A20      1
 #define INPUT_LINE_SMI      2
@@ -400,7 +398,7 @@ protected:
 	bool translate_address(int pl, int type, uint32_t *address, uint32_t *error);
 	void CHANGE_PC(uint32_t pc);
 	inline void NEAR_BRANCH(int32_t offs);
-	virtual uint8_t FETCH();
+	inline uint8_t FETCH();
 	inline uint16_t FETCH16();
 	inline uint32_t FETCH32();
 	inline uint8_t READ8(uint32_t ea) { return READ8PL(ea, m_CPL); }
@@ -1546,51 +1544,6 @@ protected:
 	virtual void WRITEPORT32(offs_t port, uint32_t value) override;
 };
 
-class i386ex_device : public i386_device
-{
-public:
-	// construction/destruction
-	i386ex_device(const machine_config &mconfig, const char *tag, device_t *owner, uint32_t clock);
-
-	uint8_t get_slave_ack(offs_t offset);
-	void set_pin_configuration(offs_t offset, uint16_t data);
-	void set_timer_configuration(offs_t offset, uint16_t data);
-	void set_serial_io_configuration(offs_t offset, uint16_t data);
-	void set_dma_configuration(offs_t offset, uint16_t data);
-	uint16_t get_interrupt_configuration(offs_t offset);
-	void set_interrupt_configuration(offs_t offset, uint16_t data);
-	void chip_select_unit_w(offs_t offset, uint16_t data);
-	void set_clock_prescaler(offs_t offset, uint16_t data);
-
-protected:
-	void io_map(address_map&);
-
-	// device-level overrides
-	virtual void device_add_mconfig(machine_config &config) override;
-	virtual void device_reset() override;
-	address_space_config m_386ex_io_config;
-
-	device_memory_interface::space_config_vector memory_space_config() const override
-	{
-	       return space_config_vector {
-		       std::make_pair(AS_PROGRAM, &m_program_config),
-		       std::make_pair(AS_IO,      &m_386ex_io_config)
-	       };
-	}
-
-private:
-	virtual uint8_t FETCH() override;
-	uint16_t m_INTCFG;
-	uint16_t m_CS_address[8];
-	uint16_t m_CS_mask[8];
-	required_device<ns16450_device> m_uart0;
-	required_device<ns16450_device> m_uart1;
-	required_device<pit8254_device> m_pit;
-	required_device<pic8259_device> m_pic_master;
-	required_device<pic8259_device> m_pic_slave;
-};
-
-
 class i486_device : public i386_device
 {
 public:
@@ -1713,7 +1666,6 @@ protected:
 
 DECLARE_DEVICE_TYPE(I386,        i386_device)
 DECLARE_DEVICE_TYPE(I386SX,      i386sx_device)
-DECLARE_DEVICE_TYPE(I386EX,      i386ex_device)
 DECLARE_DEVICE_TYPE(I486,        i486_device)
 DECLARE_DEVICE_TYPE(I486DX4,     i486dx4_device)
 DECLARE_DEVICE_TYPE(PENTIUM,     pentium_device)
